Update accounts.html
Mostly design of the page for creation of faculty accounts in account management. Looking to connect pages to gain access after admin log in. Maybe a tab, button, or drop down to chose between faculty or student. And to gain access to the account managements after being on the admin home page.